+++ public/pdf-in-page/cody-context-architecture.png
